Moved the work directory into the course repo
Before, the logs and the work was in this repo, while the course repo
was just an unmodified fork of the original course template.
That made the course repo kind of useless to have. Also, I might
work on more than just the #100DayOfWeb course during my #100DaysOfCode.

Thus, it makes more sense for me to keep the course work in the course
repo and other works in their respective repos. This repo will only
contain my logs. Which is how it was intended in the first place.