Adjust README to my personal context
Before, the README was the original #100DaysOfCode README. That was
a little weird, because it was not really focused on how I am using
this repo or the challenge.

This is fixed now and the README actually only applies to how I am
using this repo. It still contains references to the original resources.